+++ app/models/upload.rb
@@ -1,3 +1,6 @@
+require 'activerecord-import/base'
+require 'activerecord-import/active_record/adapters/postgresql_adapter'
+
 class Upload < ApplicationRecord
   belongs_to :framework, inverse_of: :uploads
 
@@ -42,7 +45,7 @@ class Upload < ApplicationRecord
     raise error if error
   end
 
-  # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
+  # rubocop:disable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity, Metrics/BlockLength
   def self.add_supplier_framework!(supplier, supplier_data)
     supplier_data[:supplier_frameworks].each do |supplier_framework_data|
       supplier_framework = supplier.supplier_frameworks.create!(supplier_framework_data.slice(:framework_id, :enabled))
@@ -53,30 +56,57 @@ class Upload < ApplicationRecord
       supplier_framework_data[:supplier_framework_lots].each do |supplier_framework_lot_data|
         supplier_framework_lot = supplier_framework.lots.create!(supplier_framework_lot_data.slice(:lot_id, :enabled))
 
-        supplier_framework_lot_data[:supplier_framework_lot_services].each do |supplier_framework_lot_service_data|
-          supplier_framework_lot.services.create!(supplier_framework_lot_service_data)
+        Supplier::Framework::Lot::Service.import!(
+          supplier_framework_lot_data[:supplier_framework_lot_services].map do |supplier_framework_lot_service_data|
+            {
+              supplier_framework_lot_id: supplier_framework_lot.id,
+              **supplier_framework_lot_service_data
+            }
+          end
+        )
+
+        supplier_framework_lot_jurisdiction_ids = Supplier::Framework::Lot::Jurisdiction.import!(
+          supplier_framework_lot_data[:supplier_framework_lot_jurisdictions].map do |supplier_framework_lot_jurisdiction_data|
+            {
+              supplier_framework_lot_id: supplier_framework_lot.id,
+              **supplier_framework_lot_jurisdiction_data
+            }
+          end
+        ).ids
+
+        jurisdiction_id_to_supplier_framework_lot_jurisdiction = supplier_framework_lot_data[:supplier_framework_lot_jurisdictions].zip(supplier_framework_lot_jurisdiction_ids).to_h do |supplier_framework_lot_jurisdiction_data, supplier_framework_lot_jurisdiction_id|
+          [
+            supplier_framework_lot_jurisdiction_data[:jurisdiction_id],
+            supplier_framework_lot_jurisdiction_id
+          ]
         end
 
-        supplier_framework_lot_data[:supplier_framework_lot_jurisdictions].each do |supplier_framework_lot_jurisdiction_data|
-          supplier_framework_lot.jurisdictions.create!(supplier_framework_lot_jurisdiction_data)
-        end
+        Supplier::Framework::Lot::Rate.import!(
+          supplier_framework_lot_data[:supplier_framework_lot_rates].map do |supplier_framework_lot_rate_data|
+            {
+              supplier_framework_lot_id: supplier_framework_lot.id,
+              supplier_framework_lot_jurisdiction_id: jurisdiction_id_to_supplier_framework_lot_jurisdiction[supplier_framework_lot_rate_data[:jurisdiction_id]],
+              **supplier_framework_lot_rate_data.except(:jurisdiction_id)
+            }
+          end
+        )
 
-        supplier_framework_lot_data[:supplier_framework_lot_rates].each do |supplier_framework_lot_rate_data|
-          supplier_framework_lot.rates.create!(supplier_framework_lot_rate_data)
-        end
-
-        supplier_framework_lot_data[:supplier_framework_lot_branches].each do |supplier_framework_lot_branches_data|
+        # Because of how slugs are generated, these need to be imported one at a time
+        supplier_framework_lot_data[:supplier_framework_lot_branches].map do |supplier_framework_lot_branches_data|
           supplier_framework_lot_branches_data[:location] = Geocoding.point(
             latitude: supplier_framework_lot_branches_data[:lat],
             longitude: supplier_framework_lot_branches_data[:lon]
           )
 
-          supplier_framework_lot.branches.create!(supplier_framework_lot_branches_data.except(:lat, :lon))
+          Supplier::Framework::Lot::Branch.create!(
+            supplier_framework_lot_id: supplier_framework_lot.id,
+            **supplier_framework_lot_branches_data.except(:lat, :lon)
+          )
         end
       end
     end
   end
-  # rubocop:enable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity
+  # rubocop:enable Metrics/AbcSize, Metrics/CyclomaticComplexity, Metrics/PerceivedComplexity, Metrics/BlockLength
 
   def self.all_or_none(framework)
     error = nil