+++ spec/helpers/legal_services/rm6240/suppliers_helper_spec.rb
@@ -2,8 +2,8 @@ require 'rails_helper'
 
 RSpec.describe LegalServices::RM6240::SuppliersHelper do
   let(:lot_number) { rand(1..3).to_s }
-  let(:jursidiction) { lot_number == '3' ? nil : ('a'..'c').to_a.sample }
-  let(:lot) { Lot.find_by(framework_id: 'RM6240', number: "#{lot_number}#{jursidiction}") }
+  let(:jurisdiction) { lot_number == '3' ? nil : ('a'..'c').to_a.sample }
+  let(:lot) { Lot.find_by(framework_id: 'RM6240', number: "#{lot_number}#{jurisdiction}") }
 
   describe '#full_lot_name' do
     it 'returns the full title with lot and description' do