+++ spec/services/management_consultancy/rm6187/admin/files_importer_spec.rb
@@ -98,9 +98,9 @@ RSpec.describe ManagementConsultancy::RM6187::Admin::FilesImporter do
   describe 'import_data' do
     let(:expected_supplier_results) do
       {
-        'REX LTD': { lots: 9, services: 131, rates: 54 },
-        'MORAG JEWEL LTD': { lots: 9, services: 131, rates: 54 },
-        'ZEKE VON GEMBU CORP': { lots: 9, services: 131, rates: 54 },
+        'REX LTD': { lots: 9, services: 131, jurisdictions: 9, rates: 54 },
+        'MORAG JEWEL LTD': { lots: 9, services: 131, jurisdictions: 9, rates: 54 },
+        'ZEKE VON GEMBU CORP': { lots: 9, services: 131, jurisdictions: 9, rates: 54 },
       }
     end
 
@@ -109,14 +109,17 @@ RSpec.describe ManagementConsultancy::RM6187::Admin::FilesImporter do
       expect(Supplier::Framework.where(framework_id: 'RM6187').count).to eq 3
     end
 
+    # rubocop:disable RSpec/MultipleExpectations
     it 'has the correct data for the suppliers' do
       expected_supplier_results.each do |name, expected_results|
         supplier_framework = Supplier.find_by(name:).supplier_frameworks.find_by(framework_id: 'RM6187')
 
         expect(supplier_framework.lots.count).to eq expected_results[:lots]
         expect(supplier_framework.lots.sum { |lot| lot.services.count }).to eq expected_results[:services]
+        expect(supplier_framework.lots.sum { |lot| lot.jurisdictions.count }).to eq expected_results[:jurisdictions]
         expect(supplier_framework.lots.sum { |lot| lot.rates.count }).to eq expected_results[:rates]
       end
     end
+    # rubocop:enable RSpec/MultipleExpectations
   end
 end